Fixed port creation bug by separating finding ports from making ports.
