+++ test/test_core_functionality.py
@@ -58,32 +58,8 @@ def test_equal():
     assert c_1 is not c_2
 
 
-class TestPort:
-    def test_in(self):
-        from BondGraphTools.base import Port
-        from BondGraphTools.actions import new
 
-        c = new("C")
-        one = new("1")
-
-        port_c = Port(c, 0)
-
-        assert c in port_c
-        assert c is not port_c
-        assert one is not port_c
-        assert one not in port_c
-
-
-
-    def test_cmp(self):
-        c = new("C")
-        one = new("1")
-
-        port_c = Port(c, 0)
-        port_one = Port(one, 0)
-
-        assert port_c is not port_one
-        assert port_c != port_one
+#    def test_get_exposed_port
 
 
 class TestBond:
@@ -101,6 +77,17 @@ class TestBond:
         assert (c, 0) in b_1
         assert (one, 0) in b_1
         assert (one, 1) not in b_1
+
+    def test_compare(self):
+        from BondGraphTools.base import Bond
+        from BondGraphTools.actions import new
+
+        c = new('C')
+        one = new('1')
+
+        b_1 = Bond(head=(c, 0), tail=(one, 0))
+
+        assert b_1 == ((one, 0), (c, 0))
 
 
 class TestConnect:
@@ -195,6 +182,8 @@ class TestConnect:
                     assert component in bond
                 else:
                     assert component not in bond
+
+
 
 def test_disconnect_ports():
 
@@ -390,12 +379,12 @@ class TestSwap:
         assert c in bg.components
         r_p, = r.ports
         c_p, = c.ports
-        z0,z1, = zero.ports
+        z0, z1, = zero.ports
 
-        assert set(bg.bonds) == {
+        assert bg.bonds == [
             (r_p, z0),
             (c_p, z1)
-        }
+        ]
 
         assert len(bg.state_vars) == 1
         Sf = bgt.new('Sf')
@@ -404,20 +393,20 @@ class TestSwap:
         assert len(bg.state_vars) == 0
         assert len(bg.control_vars) == 1
 
-        assert set(bg.bonds) == {
+        assert bg.bonds == [
             (r_p, z0),
             (sf_port, z1)
-        }
+        ]
 
         assert c not in bg.components
         assert Sf in bg.components
 
         swap(Sf, c)
 
-        assert set(bg.bonds) == {
+        assert bg.bonds == [
             (r_p, z0),
             (c_p, z1)
-        }
+        ]
 
         assert c in bg.components
         assert Sf not in bg.components
@@ -456,7 +445,8 @@ class TestSwap:
         connect(r, zero)
         connect(c, zero)
 
-        with pytest.raises(InvalidComponentException):
+        with pytest.raises(InvalidPortException):
             swap(zero, l)
+
